website-django/pyproject.toml
[tool.poetry]
name = "website-django"
version = "0.1.0"
description = ""
authors = ["Tobias Schulmann"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.1.5"
wagtail = "^6.3.2"
crowdin-api-client = "^1.19.0"
psycopg = {extras = ["binary"], version = "^3.2.4"}
wagtail-bakery = { git = "https://github.com/Stormheg/wagtail-bakery.git", rev = "a491649a2bee8790c50651e3e9756f2a73b28d71"}
redis = {extras = ["hiredis"], version = "^5.2.1"}
django-rq = "^3.0.0"
wagtail-localize = "^1.11.1"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
sentry-sdk = {extras = ["django"], version = "^2.22.0"}

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"
ruff = "^0.9.2"
djlint = "^1.36.4"
djhtml = "^3.0.7"
django-browser-reload = "^1.17.0"
requests = "^2.32.3"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
poethepoet = "^0.32.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
extend-exclude = ["*/migrations/*", "taska_schemas/*"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "UP", "S", "A", "I", "T20"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = ["S101"]

[tool.djlint]
profile = "django"
indent = 2
max_line_length = 100
ignore="H037"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "website_django.settings.dev"
addopts = "--reuse-db"

[tool.poe.tasks]
build = "python manage.py build --settings website_django.settings.production"